refactor(dashboard): extract WikiSpriteService from DevServer
Move the wiki sprite/texture serving cluster (16 methods, ~440 lines: spritesheet.xml parsing, atlas sprite cropping via sharp, case-insensitive PNG resolution, extractor game-data dir resolution) plus its types/constant and the wikiSpriteSheetCache into src/dashboard/server/wikiSpriteService.ts. DevServer injects 3 deps (publicDir, getRotmgPath, getExtractorGameDataPath) and delegates the /api/wiki-texture-file route + cache reset. DevServer.ts: 5,168 -> 4,719. tsc clean.
